requirements.txt
aio-pika==9.4.3
aiormq==6.8.0
aiosqlite==0.17.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
idna==3.8
iso8601==1.1.0
multidict==6.0.5
pamqp==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
tortoise-orm==0.21.6
typing_extensions==4.12.2
yarl==1.9.4
